+++ examples/04_environment_node.rs
@@ -11,7 +11,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     environment.set_listener_position(1.0, 2.0, -3.0);
     environment.set_listener_orientation(15.0, 5.0, 0.0);
-    environment.set_distance_attenuation(1, 1.0, 25.0, 0.7);
+    environment.set_distance_attenuation(1, 1.0, 25.0, 0.7)?;
     environment.set_reverb_blend(20.0);
 
     println!("listener position: {:?}", environment.listener_position()?);